lesson12

+++ Core-Service/src/main/java/ru/geekbrains/web/core/repositories/specifications/ProductSpecifications.java
@@ -4,12 +4,14 @@ package ru.geekbrains.web.core.repositories.specifications;
 import org.springframework.data.jpa.domain.Specification;
 import ru.geekbrains.web.core.model.Product;
 
+import java.math.BigDecimal;
+
 public class ProductSpecifications {
-    public static Specification<Product> priceGreaterOrEqualsThan(int minPrice) {
+    public static Specification<Product> priceGreaterOrEqualsThan(BigDecimal minPrice) {
         return (root, criteriaQuery, criteriaBuilder) -> criteriaBuilder.greaterThanOrEqualTo(root.get("price"), minPrice);
     }
 
-    public static Specification<Product> priceLesserOrEqualsThan(int maxPrice) {
+    public static Specification<Product> priceLesserOrEqualsThan(BigDecimal maxPrice) {
         return (root, criteriaQuery, criteriaBuilder) -> criteriaBuilder.lessThanOrEqualTo(root.get("price"), maxPrice);
     }
